feat: if DB empty, add basic data and duplicate prevention
